taxinnovation/apps/endpoint_pub/requirements.txt
gunicorn==20.0.4
google-cloud-pubsub==2.1.0
google-cloud-logging==1.15.1
google-auth==1.22.1
Flask==1.1.2
psycopg2-binary==2.8.6